pyproject.toml
[project]
name = "switchbase-teamview"
version = "0.1.0"
description = "Python client and CLI for TeamView management APIs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "lark-oapi>=1.5.3",
    "matplot>=0.1.9",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
    "python-socks>=2.8.1",
    "typer>=0.16.0",
]

[project.scripts]
teamview-cli = "switchbase_teamview.cli:run"
teamview-api = "switchbase_teamview.api:main"
teamview-report-daemon = "switchbase_teamview.report_daemon:main"
teamview-feishu-bot = "switchbase_teamview.feishu_bot_main:main"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["switchbase_teamview", "scripts"]

[tool.hatch.build.targets.sdist]
include = [
    ".codex/check-maxline.json",
    "switchbase_teamview/**/*.py",
    "scripts/**/*.py",
    "tests/**/*.py",
    ".env.template",
    ".gitignore",
    "README.md",
    "teamview_aliases.template.json",
    "pyproject.toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
